+++ example/lib/media_stream_extension.dart
@@ -2,7 +2,9 @@ import 'package:flutter_webrtc/flutter_webrtc.dart';
 
 extension MediaStreamExtension on MediaStream {
   Future<void> disposeTracks() async => await Future.wait(
-        getTracks().map((MediaStreamTrack track) => track.stop()).toList(growable: false),
+        getTracks()
+            .map((MediaStreamTrack track) => track.stop())
+            .toList(growable: false),
       );
 
   Future<void> disposeMediaStream() async {