+++ lib/flutter_aira.dart
@@ -1,27 +1,40 @@
 export 'src/enum/user_property.dart' show UserProperty;
 export 'src/messaging_client.dart' show MessagingClient;
-export 'src/models/access_offer.dart' show AccessOfferDetails, AccessOfferType, AccountDetails, SiteAddress;
+export 'src/models/access_offer.dart'
+    show AccessOfferDetails, AccessOfferType, AccountDetails, SiteAddress;
 export 'src/models/account.dart' show Account, AccountType;
 export 'src/models/billing_info.dart' show PartialBillingInformation;
-export 'src/models/call_history.dart' show CallSession, SessionFeedback;
-export 'src/models/chat/chat.dart' show ChatMessageInfo, ChatSessionInfo, SenderRole;
+export 'src/models/callHistory/build_ai.dart'
+    show BuildAi, BuildAiStatus, NotAllowSharingReason;
+export 'src/models/callHistory/call_history.dart'
+    show CallSession, SessionFeedback;
+export 'src/models/chat/agent_feedback_info.dart'
+    show AgentFeedbackInfo, AgentFeedbackState;
+export 'src/models/chat/chat.dart'
+    show ChatMessageInfo, ChatSessionInfo, SenderRole;
 export 'src/models/chat/user_feedback_info.dart' show UserFeedbackInfo;
-export 'src/models/chat/agent_feedback_info.dart' show AgentFeedbackInfo, AgentFeedbackState;
 export 'src/models/credentials.dart' show Credentials;
 export 'src/models/feedback.dart' show AgentFeedback, Feedback, Rating;
 export 'src/models/message.dart' show Message;
-export 'src/models/minute_sharing.dart' show MinuteSharingMember, MinuteSharingInformation;
+export 'src/models/minute_sharing.dart'
+    show MinuteSharingMember, MinuteSharingInformation;
 export 'src/models/paged.dart' show Paged;
 export 'src/models/participant.dart' show Participant, ParticipantStatus;
 export 'src/models/photo.dart' show Photo;
 export 'src/models/position.dart' show Position;
 export 'src/models/profile.dart' show Language, Profile, ProfileType;
-export 'src/models/service_request.dart' show ServiceRequest, ServiceRequestState;
+export 'src/models/service_request.dart'
+    show ServiceRequest, ServiceRequestState;
 export 'src/models/session.dart' show Session;
 export 'src/models/track.dart' show Track;
 export 'src/models/usage.dart' show Usage, PlanUsageBreakdown;
-export 'src/models/user.dart' show User;
-export 'src/platform_client.dart' show PlatformClient, PlatformClientConfig, PlatformEnvironment, PlatformMessagingKeys;
+export 'src/models/user.dart' show User, AiLanguageLevel, AiVerbosity;
+export 'src/platform_client.dart'
+    show
+        PlatformClient,
+        PlatformClientConfig,
+        PlatformEnvironment,
+        PlatformMessagingKeys;
 export 'src/platform_exceptions.dart'
     show
         PlatformDeleteAccountException,