+++ lib/src/models/participant_message.dart
@@ -1,4 +1,6 @@
 // ignore_for_file: constant_identifier_names
+import 'package:flutter_aira/src/models/track.dart';
+
 enum ParticipantMessageType {
   ICE_CANDIDATE,
   INCOMING_TRACK_CREATE,
@@ -11,8 +13,11 @@ enum ParticipantMessageType {
 extension ParticipantMessageTypeExtension on ParticipantMessageType {
   get name => toString().split('.').last;
 
-  static ParticipantMessageType fromName(String name) => ParticipantMessageType.values
-      .firstWhere((type) => type.name == name, orElse: () => throw UnimplementedError(name));
+  static ParticipantMessageType fromName(String name) =>
+      ParticipantMessageType.values.firstWhere(
+        (type) => type.name == name,
+        orElse: () => throw UnimplementedError(name),
+      );
 }
 
 class ParticipantMessage {
@@ -30,6 +35,14 @@ class ParticipantMessage {
           json['participantId'],
           json['payload'],
         );
+
+  TrackKind? get trackKind => (isAudio ?? false
+      ? TrackKind.audio
+      : (isVideo ?? false ? TrackKind.video : null));
+
+  bool? get isAudio => bool.tryParse(payload['audio'].toString());
+
+  bool? get isVideo => bool.tryParse(payload['video'].toString());
 
   Map<String, dynamic> toJson() {
     return {