+++ lib/src/models/sent_file_info.dart
@@ -1,4 +1,3 @@
-
 /// Implementation independent information on sent file.
 class SentFileInfo {
   SentFileInfo(this.id, {this.url, this.name});